refactor(settings): extract Networks Join dialog into its own component (PR-U3 slice 5)
Second structural slice: pull the Join dialog out of NetworksComponent into NetworkJoinDialogComponent.
It owns the invite-bundle textarea, the bundle validation (JSON / required fields / this brain's URL),
the space-id collision-resolution UI (merge vs alias, with alias validation), and the joinRemote call.
`myUrl` is a one-way input (the host computes this brain's own URL; the join dialog never edits it), and
a successful join emits `joined` so the host reloads networks and refreshes its spaces list.

Behavior unchanged. The join characterization tests moved to network-join-dialog.component.spec.ts
(bundle validation, collision hold, alias validation in confirmJoin, immediate join + `joined` emit); the
parent keeps a test that onJoined reloads. Parent has now shed ~250 lines across the two extractions.
28 networks specs pass; AOT build clean.
